requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-bigquery==3.20.1
google-auth==2.29.0
